requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1

# LLM
openai==1.59.4

# Google Sheets
gspread==6.1.4
google-auth==2.38.0

# Slack
slack-sdk==3.34.0
slack-bolt==1.22.0

# Observability
loguru==0.7.3
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Dev
ruff==0.8.6
